+++ src/components/SortSearchResults/SortOptions.tsx
@@ -1,13 +1,11 @@
+import { Option } from '@components/SortSearchResults/SearchResultSorter'
 import React, { useState } from 'react'
 import { Pressable, StyleSheet, Text, View } from 'react-native'
-import { SortOptions as SortOptionsType } from '@futureproof/typings'
 
 interface ComponentProps {
-  options: { label: string; value: SortOptionsType }[]
+  options: Option[]
   initial?: number | null
-  onChange: React.Dispatch<
-    React.SetStateAction<{ label: string; value: SortOptionsType }>
-  >
+  onChange: React.Dispatch<React.SetStateAction<Option>>
 }
 
 /**
@@ -29,17 +27,9 @@ const SortOptions = ({ options, initial, onChange }: ComponentProps) => {
       initial >= 0 && initial < options.length ? options[initial] : null
   }
 
-  // const io = !initial ? ((initial >= 0 && initial < options.length) ? options[initial] : null) : null
+  const [userOption, setUserOption] = useState<Option | null>(initialOption)
 
-  const [userOption, setUserOption] = useState<{
-    label: string
-    value: SortOptionsType
-  } | null>(initialOption)
-
-  const updateOptionChoice = (option: {
-    label: string
-    value: SortOptionsType
-  }) => {
+  const updateOptionChoice = (option: Option) => {
     onChange(option)
     setUserOption(option)
   }
@@ -50,17 +40,16 @@ const SortOptions = ({ options, initial, onChange }: ComponentProps) => {
         return (
           <Pressable
             key={option.value}
-            style={({ pressed }) => [
-              {
-                backgroundColor: pressed ? '#ededed' : '#ffffff',
-                opacity: pressed ? 0.5 : 1,
-              },
-              styles.option,
-              option === userOption
+            style={({ pressed }) => ({
+              backgroundColor: pressed ? '#ededed' : '#ffffff',
+              opacity: pressed ? 0.5 : 1,
+              ...(option === userOption
                 ? styles.selected_option
-                : styles.unselected_option,
-            ]}
+                : styles.unselected_option),
+              ...styles.option,
+            })}
             onPress={() => updateOptionChoice(option)}
+            testID={`sort-option-${option.value}`}
           >
             <Text
               style={[
