+++ src/components/ratings/CircularRatingIndicator.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 import Svg, { Circle } from 'react-native-svg'
-import { ratingToColour } from '@components/search/RatingCapsule'
+import { ratingToColour } from '@components/Search/RatingCapsule'
 
 const BACKGROUND_SECONDARY_COLOUR = '#FAF9F9'
