+++ src/components/ratings/RectangularRatingIndicator.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 import Svg, { Rect } from 'react-native-svg'
-import { ratingToColour } from '@components/search/RatingCapsule'
+import { ratingToColour } from '@components/Search/RatingCapsule'
 
 const PROGRESS_BACKGROUND_COLOUR = '#F0EFEF'
